Added C API, JS/WASM API
The C API will be the public interface to Qak. Sadly, I have to shuffle around data from the internal C++ representation to the C structs exposed through the API. However, in the usual case things like tokens or AST nodes will not be queried by the user, so this doesn't have to be super high performance.

The JS/WASM API exposes the C interface in an idiomatic JS way.